cleaned up lots of code
