+++ src/actions/board.js
@@ -3,6 +3,6 @@ export const BOARD = 'BOARD'
 export function setBoard(shipArr) {
   return {
     type: BOARD,
-    board: [shipArr],
+    board: shipArr,
   }
 }
